# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Прогноз отказов оборудования коллекторов",
    "description": "REST API сервиса прогноза отказов датчиков коллекторов на 24 часа (ЛЦТ-2026).\n\n**Вход:** нажмите *Authorize* и введите логин и пароль (в демо-режиме — демо-пользователя из README), поля client_id/secret оставьте пустыми.\n\n**Машина времени:** все запросы на чтение принимают `at`. Демо-данные покрывают 01.07.2026–31.08.2026, например `at=2026-08-01T12:00:00`.\n\n**Симуляция потока** (`/api/sim/*`): проигрывание суток по заранее рассчитанным прогнозам с ускорением; пока она идёт, запросы без `at` работают на модельное время. Онлайн-пересчёт моделью не выполняется.\n\nВремя — московское, без часового пояса.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/work-orders": {
      "post": {
        "tags": [
          "Заявки"
        ],
        "summary": "Создать черновик заявки (из прогноза — с автозаполнением)",
        "description": "Из `prediction_id` подставляются датчик, объект и — по рекомендации движка правил ТО (`GET /api/predictions/{id}/recommendation`) — приоритет, срок, исполнитель, рекомендация и описание с обоснованием и причинами прогноза. Любое поле можно передать явно. `at` в запросе — момент «машины времени», которым датируется заявка.",
        "operationId": "create_work_order_api_work_orders_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "at",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Момент «машины времени» (ISO 8601). Без пояса — московское время. Не задан — текущее время, а во время симуляции — модельное время.",
              "examples": [
                "2026-08-01T12:00:00"
              ],
              "title": "At"
            },
            "description": "Момент «машины времени» (ISO 8601). Без пояса — московское время. Не задан — текущее время, а во время симуляции — модельное время."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkOrderIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkOrderOut"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Заявки"
        ],
        "summary": "Список заявок с фильтрами",
        "operationId": "list_work_orders_api_work_orders_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "enum": [
                      "draft",
                      "submitted",
                      "in_progress",
                      "done",
                      "cancelled"
                    ],
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "Можно несколько",
              "title": "Status"
            },
            "description": "Можно несколько"
          },
          {
            "name": "priority",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "enum": [
                      "low",
                      "medium",
                      "high",
                      "critical"
                    ],
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Priority"
            }
          },
          {
            "name": "object_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Объект; включает все вложенные",
              "title": "Object Id"
            },
            "description": "Объект; включает все вложенные"
          },
          {
            "name": "channel_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Channel Id"
            }
          },
          {
            "name": "prediction_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Prediction Id"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Поиск по номеру, датчику, исполнителю",
              "title": "Q"
            },
            "description": "Поиск по номеру, датчику, исполнителю"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "Размер страницы",
              "default": 50,
              "title": "Limit"
            },
            "description": "Размер страницы"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Сдвиг от начала",
              "default": 0,
              "title": "Offset"
            },
            "description": "Сдвиг от начала"
          },
          {
            "name": "at",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Момент «машины времени» (ISO 8601). Без пояса — московское время. Не задан — текущее время, а во время симуляции — модельное время.",
              "examples": [
                "2026-08-01T12:00:00"
              ],
              "title": "At"
            },
            "description": "Момент «машины времени» (ISO 8601). Без пояса — московское время. Не задан — текущее время, а во время симуляции — модельное время."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkOrderList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/maintenance/plan": {
      "get": {
        "tags": [
          "Рекомендации по ТО"
        ],
        "summary": "Рекомендованные работы: профилактика на ближайшие дни по правилам ТО",
        "description": "Датчики среза, которым по правилам нужна профилактика (повторные отказы, повтор после ремонта, деградация показаний, агрегаты, питание и связь объекта) и по которым ещё нет открытой заявки. Сгруппированы по объектам: сначала объекты с самым высоким приоритетом. Кандидаты — прогнозы «Внимание» и выше и датчики с 2+ отказами за 30 дней.",
        "operationId": "plan_api_maintenance_plan_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "object_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Объект; включает вложенные",
              "title": "Object Id"
            },
            "description": "Объект; включает вложенные"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "Объектов на странице",
              "default": 50,
              "title": "Limit"
            },
            "description": "Объектов на странице"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "at",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Момент «машины времени» (ISO 8601). Без пояса — московское время. Не задан — текущее время, а во время симуляции — модельное время.",
              "examples": [
                "2026-08-01T12:00:00"
              ],
              "title": "At"
            },
            "description": "Момент «машины времени» (ISO 8601). Без пояса — московское время. Не задан — текущее время, а во время симуляции — модельное время."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaintenancePlan"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/maintenance/drafts": {
      "post": {
        "tags": [
          "Рекомендации по ТО"
        ],
        "summary": "Создать черновики заявок по рекомендациям (выбранным прогнозам)",
        "description": "Для каждого прогноза — черновик по рекомендации ТО, как `POST /api/work-orders` с `prediction_id`. Пропускаются: датчики с открытой заявкой, рекомендации без заявки (норма, наблюдение). Для правил «на объект» (массовая потеря связи, питание объекта) — одна заявка на объект, остальные датчики перечисляются в её описании. Каждая заявка пишется в `audit_log`.",
        "operationId": "drafts_api_maintenance_drafts_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "at",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Момент «машины времени» (ISO 8601). Без пояса — московское время. Не задан — текущее время, а во время симуляции — модельное время.",
              "examples": [
                "2026-08-01T12:00:00"
              ],
              "title": "At"
            },
            "description": "Момент «машины времени» (ISO 8601). Без пояса — московское время. Не задан — текущее время, а во время симуляции — модельное время."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MaintenanceDraftsIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaintenanceDraftsOut"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChannelRef": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id",
            "examples": [
              334609
            ]
          },
          "name": {
            "type": "string",
            "title": "Name",
            "examples": [
              "Дым ПК 1101+2"
            ]
          },
          "sensor_type": {
            "type": "string",
            "title": "Sensor Type",
            "examples": [
              "Датчик дыма"
            ]
          },
          "system_type": {
            "type": "string",
            "title": "System Type",
            "examples": [
              "Пожарная охрана"
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "sensor_type",
          "system_type"
        ],
        "title": "ChannelRef"
      },
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "examples": [
              "Прогноз не найден"
            ]
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MaintenanceDraftsIn": {
        "properties": {
          "prediction_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "maxItems": 100,
            "minItems": 1,
            "title": "Prediction Ids",
            "description": "Прогнозы из «Рекомендованных работ»"
          }
        },
        "type": "object",
        "required": [
          "prediction_ids"
        ],
        "title": "MaintenanceDraftsIn"
      },
      "MaintenanceDraftsOut": {
        "properties": {
          "created": {
            "items": {
              "$ref": "#/components/schemas/WorkOrderOut"
            },
            "type": "array",
            "title": "Created"
          },
          "skipped": {
            "items": {
              "$ref": "#/components/schemas/MaintenanceDraftSkipped"
            },
            "type": "array",
            "title": "Skipped"
          }
        },
        "type": "object",
        "required": [
          "created",
          "skipped"
        ],
        "title": "MaintenanceDraftsOut"
      },
      "MaintenancePlan": {
        "properties": {
          "at": {
            "type": "string",
            "format": "date-time",
            "title": "At"
          },
          "snapshot_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snapshot At"
          },
          "total_items": {
            "type": "integer",
            "title": "Total Items"
          },
          "total_objects": {
            "type": "integer",
            "title": "Total Objects"
          },
          "groups": {
            "items": {
              "$ref": "#/components/schemas/MaintenancePlanGroup"
            },
            "type": "array",
            "title": "Groups"
          },
          "source": {
            "type": "string",
            "title": "Source"
          }
        },
        "type": "object",
        "required": [
          "at",
          "snapshot_at",
          "total_items",
          "total_objects",
          "groups",
          "source"
        ],
        "title": "MaintenancePlan"
      },
...
      "WorkOrderIn": {
        "properties": {
          "prediction_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prediction Id",
            "description": "Из какого прогноза создаётся черновик — поля заполнятся сами"
          },
          "channel_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channel Id",
            "description": "Нужен, если заявка не из прогноза"
          },
          "reason_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason Id"
          },
          "recommendation_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommendation Id"
          },
          "recommendation_text": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4000
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommendation Text"
          },
          "priority": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "low",
                  "medium",
                  "high",
                  "critical"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority",
            "description": "По умолчанию — из уровня риска"
          },
          "due_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due At",
            "description": "По умолчанию — из приоритета"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "assignee": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignee"
          }
        },
        "type": "object",
        "title": "WorkOrderIn"
      },
      "WorkOrderList": {
        "properties": {
          "at": {
            "type": "string",
            "format": "date-time",
            "title": "At"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/WorkOrderOut"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "at",
          "total",
          "items"
        ],
        "title": "WorkOrderList"
      },
      "WorkOrderOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "number": {
            "type": "string",
            "title": "Number",
            "examples": [
              "ЗН-2026-000042"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "submitted",
              "in_progress",
              "done",
              "cancelled"
            ],
            "title": "Status"
          },
          "status_label": {
            "type": "string",
            "title": "Status Label"
          },
          "priority": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high",
              "critical"
            ],
            "title": "Priority"
          },
          "priority_label": {
            "type": "string",
            "title": "Priority Label"
          },
          "prediction_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prediction Id"
          },
          "channel": {
            "$ref": "#/components/schemas/ChannelRef"
          },
          "object": {
            "$ref": "#/components/schemas/ObjectRef"
          },
          "reason": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReasonRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "recommendation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RecommendationRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "recommendation_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommendation Text"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "assignee": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignee"
          },
          "due_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due At"
          },
          "created_by": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "closed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "number",
          "status",
          "status_label",
          "priority",
          "priority_label",
          "prediction_id",
          "channel",
          "object",
          "reason",
          "recommendation",
          "recommendation_text",
          "description",
          "assignee",
          "due_at",
          "created_by",
          "created_at",
          "updated_at",
          "closed_at"
        ],
        "title": "WorkOrderOut"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/token"
          }
        }
      }
    }
  }
}